+++ crates/symsorter/src/app.rs
@@ -7,13 +7,10 @@ use std::sync::Mutex;
 use anyhow::{anyhow, Context, Result};
 use chrono::{DateTime, Utc};
 use console::style;
-use object::File;
-// use object::{macho::DyldCacheHeader};
-// use object::macho::DyldCacheMappingInfo;
-use object::{
-    read::macho::DyldCache, write, Endianness, Object, ObjectComdat, ObjectSection, ObjectSymbol,
-    RelocationTarget, SectionKind, SymbolFlags, SymbolKind, SymbolSection,
-};
+use object::macho::MachHeader32;
+use object::macho::MachHeader64;
+use object::read::macho::{DyldCache, DyldCacheImage, MachHeader};
+use object::{Endianness, Object};
 use rayon::prelude::*;
 use serde::Serialize;
 use structopt::StructOpt;
@@ -25,7 +22,8 @@ use zstd::stream::copy_encode;
 
 use crate::config::{RunConfig, SortConfig};
 use crate::utils::{
-    create_source_bundle, get_target_filename, get_unified_id, is_bundle_id, make_bundle_id,
+    clone_dyld_image, create_source_bundle, get_dyld_image_filename, get_dyld_image_unified_id,
+    get_target_filename, get_unified_id, is_bundle_id, make_bundle_id,
 };
 
 /// Sorts debug symbols into the right structure for symbolicator.
@@ -173,148 +171,12 @@ fn process_file(
     Ok(rv)
 }
 
-fn clone_dyld_image(in_object: &File) -> Vec<u8> {
-    let mut out_object = write::Object::new(
-        in_object.format(),
-        in_object.architecture(),
-        in_object.endianness(),
-    );
-    out_object.mangling = write::Mangling::None;
-    out_object.flags = in_object.flags();
-
-    let mut out_sections = HashMap::new();
-    for in_section in in_object.sections() {
-        if in_section.kind() == SectionKind::Metadata {
-            continue;
-        }
-        let section_id = out_object.add_section(
-            in_section
-                .segment_name()
-                .unwrap()
-                .unwrap_or("")
-                .as_bytes()
-                .to_vec(),
-            in_section.name().unwrap().as_bytes().to_vec(),
-            in_section.kind(),
-        );
-        let out_section = out_object.section_mut(section_id);
-        if out_section.is_bss() {
-            out_section.append_bss(in_section.size(), in_section.align());
-        } else {
-            out_section.set_data(in_section.data().unwrap().into(), in_section.align());
-        }
-        out_section.flags = in_section.flags();
-        out_sections.insert(in_section.index(), section_id);
-    }
-
-    let mut out_symbols = HashMap::new();
-    for in_symbol in in_object.symbols() {
-        match in_symbol.kind() {
-            SymbolKind::Unknown => continue,
-            // null and label are unsupported according to macho_write()
-            SymbolKind::Null => continue,
-            SymbolKind::Label => continue,
-            _ => (),
-        }
-        let (section, value) = match in_symbol.section() {
-            SymbolSection::None => (write::SymbolSection::None, in_symbol.address()),
-            SymbolSection::Undefined => (write::SymbolSection::Undefined, in_symbol.address()),
-            SymbolSection::Absolute => (write::SymbolSection::Absolute, in_symbol.address()),
-            SymbolSection::Common => (write::SymbolSection::Common, in_symbol.address()),
-            SymbolSection::Section(index) => {
-                if let Some(out_section) = out_sections.get(&index) {
-                    (
-                        write::SymbolSection::Section(*out_section),
-                        in_symbol.address() - in_object.section_by_index(index).unwrap().address(),
-                    )
-                } else {
-                    // Ignore symbols for sections that we have skipped.
-                    assert_eq!(in_symbol.kind(), SymbolKind::Section);
-                    continue;
-                }
-            }
-            _ => panic!("unknown symbol section for {:?}", in_symbol),
-        };
-        let flags = match in_symbol.flags() {
-            SymbolFlags::None => SymbolFlags::None,
-            SymbolFlags::Elf { st_info, st_other } => SymbolFlags::Elf { st_info, st_other },
-            SymbolFlags::MachO { n_desc } => SymbolFlags::MachO { n_desc },
-            SymbolFlags::CoffSection {
-                selection,
-                associative_section,
-            } => {
-                let associative_section =
-                    associative_section.map(|index| *out_sections.get(&index).unwrap());
-                SymbolFlags::CoffSection {
-                    selection,
-                    associative_section,
-                }
-            }
-            _ => panic!("unknown symbol flags for {:?}", in_symbol),
-        };
-        let out_symbol = write::Symbol {
-            name: in_symbol.name().unwrap_or("").as_bytes().to_vec(),
-            value,
-            size: in_symbol.size(),
-            kind: in_symbol.kind(),
-            scope: in_symbol.scope(),
-            weak: in_symbol.is_weak(),
-            section,
-            flags,
-        };
-        let symbol_id = out_object.add_symbol(out_symbol);
-        out_symbols.insert(in_symbol.index(), symbol_id);
-    }
-
-    for in_section in in_object.sections() {
-        if in_section.kind() == SectionKind::Metadata {
-            continue;
-        }
-        let out_section = *out_sections.get(&in_section.index()).unwrap();
-        for (offset, in_relocation) in in_section.relocations() {
-            let symbol = match in_relocation.target() {
-                RelocationTarget::Symbol(symbol) => *out_symbols.get(&symbol).unwrap(),
-                RelocationTarget::Section(section) => {
-                    out_object.section_symbol(*out_sections.get(&section).unwrap())
-                }
-                _ => panic!("unknown relocation target for {:?}", in_relocation),
-            };
-            let out_relocation = write::Relocation {
-                offset,
-                size: in_relocation.size(),
-                kind: in_relocation.kind(),
-                encoding: in_relocation.encoding(),
-                symbol,
-                addend: in_relocation.addend(),
-            };
-            out_object
-                .add_relocation(out_section, out_relocation)
-                .unwrap();
-        }
-    }
-
-    for in_comdat in in_object.comdats() {
-        let mut sections = Vec::new();
-        for in_section in in_comdat.sections() {
-            sections.push(*out_sections.get(&in_section).unwrap());
-        }
-        out_object.add_comdat(write::Comdat {
-            kind: in_comdat.kind(),
-            symbol: *out_symbols.get(&in_comdat.symbol()).unwrap(),
-            sections,
-        });
-    }
-
-    out_object.write().unwrap()
-}
-
-fn process_dyld_file(
+fn process_dyld_image(
     sort_config: &SortConfig,
-    bv: ByteView,
-    writable_object: object::File,
-    filename: String,
+    image: DyldCacheImage,
 ) -> Result<Vec<(String, ObjectKind)>> {
     let mut rv = vec![];
+    let filename = image.path()?.rsplit('/').next().unwrap().to_string();
 
     macro_rules! maybe_ignore_error {
         ($expr:expr) => {
@@ -324,11 +186,9 @@ fn process_dyld_file(
                     if RunConfig::get().ignore_errors {
                         eprintln!(
                             "{}: ignored error {} ({})",
-                            "error",
-                            // style("error").red().bold(),
+                            style("error").red().bold(),
                             err,
-                            filename,
-                            // style(filename).cyan(),
+                            style(filename).cyan(),
                         );
 
                         for cause in err.chain().skip(1) {
@@ -350,53 +210,84 @@ fn process_dyld_file(
         3 => 19,
         _ => 22,
     };
-    let archive = maybe_ignore_error!(
-        Archive::parse(&bv).map_err(|e| anyhow!("failed to parse archive {}", e))
-    );
+
+    let in_object = maybe_ignore_error!(image
+        .parse_object()
+        .map_err(|e| anyhow!("failed to parse archive {}", e)));
     let root = &RunConfig::get().output;
 
-    for obj in archive.objects() {
-        let obj = maybe_ignore_error!(obj.map_err(|e| anyhow!(e)));
-        let new_filename = root.join(maybe_ignore_error!(
-            get_target_filename(&obj).ok_or_else(|| anyhow!("unsupported file"))
-        ));
+    let writable_data = clone_dyld_image(&in_object);
 
-        fs::create_dir_all(new_filename.parent().unwrap())?;
+    let raw_kind = if in_object.is_64() {
+        MachHeader64::parse(writable_data.as_slice(), 0)?.filetype(in_object.endianness())
+    } else {
+        MachHeader32::parse(writable_data.as_slice(), 0)?.filetype(in_object.endianness())
+    };
+    let kind = match raw_kind {
+        object::macho::MH_OBJECT => ObjectKind::Relocatable,
+        object::macho::MH_EXECUTE => ObjectKind::Executable,
+        object::macho::MH_FVMLIB => ObjectKind::Library,
+        object::macho::MH_CORE => ObjectKind::Dump,
+        object::macho::MH_PRELOAD => ObjectKind::Executable,
+        object::macho::MH_DYLIB => ObjectKind::Library,
+        object::macho::MH_DYLINKER => ObjectKind::Executable,
+        object::macho::MH_BUNDLE => ObjectKind::Library,
+        object::macho::MH_DSYM => ObjectKind::Debug,
+        object::macho::MH_KEXT_BUNDLE => ObjectKind::Library,
+        _ => ObjectKind::Other,
+    };
 
-        if let Some(bundle_id) = &sort_config.bundle_id {
-            let refs_path = new_filename.parent().unwrap().join("refs");
-            fs::create_dir_all(&refs_path)?;
-            fs::write(&refs_path.join(bundle_id), b"")?;
-        }
+    let new_filename = root.join(maybe_ignore_error!(get_dyld_image_filename(
+        &in_object, &kind
+    )
+    .ok_or_else(|| anyhow!("unsupported file"))));
 
-        let meta = DebugIdMeta {
-            name: Some(filename.clone()),
-            arch: Some(obj.arch()),
-            file_format: Some(obj.file_format()),
-        };
+    fs::create_dir_all(new_filename.parent().unwrap())?;
 
-        fs::write(
-            &new_filename.parent().unwrap().join("meta"),
-            &serde_json::to_vec(&meta)?,
-        )?;
-
-        log!(
-            "{} ({}, {}) -> {}",
-            style(&filename).dim(),
-            style(obj.kind()).yellow(),
-            style(obj.arch()).yellow(),
-            style(new_filename.display()).cyan(),
-        );
-        let mut out = fs::File::create(&new_filename)?;
-
-        let writable_data = clone_dyld_image(&writable_object);
-        if compression_level > 0 {
-            copy_encode(writable_data.as_slice(), &mut out, compression_level)?;
-        } else {
-            io::copy(&mut writable_data.as_slice(), &mut out)?;
-        }
-        rv.push((get_unified_id(&obj), obj.kind()));
+    if let Some(bundle_id) = &sort_config.bundle_id {
+        let refs_path = new_filename.parent().unwrap().join("refs");
+        fs::create_dir_all(&refs_path)?;
+        fs::write(&refs_path.join(bundle_id), b"")?;
     }
+
+    // all of the types recognized by DyldCacheHeader
+    let arch = match in_object.architecture() {
+        object::Architecture::I386 => Arch::X86,
+        object::Architecture::PowerPc => Arch::Ppc,
+        // we lose a bit of precision in the below variants because object combines all x86_64, arm,
+        // and arm64 variants into one common enum member
+        object::Architecture::X86_64 => Arch::Amd64,
+        object::Architecture::Arm => Arch::Arm,
+        object::Architecture::Aarch64 => Arch::Arm64_32,
+        _ => Arch::Unknown,
+    };
+
+    let meta = DebugIdMeta {
+        name: Some(filename.clone()),
+        arch: Some(arch),
+        file_format: Some(FileFormat::MachO),
+    };
+
+    fs::write(
+        &new_filename.parent().unwrap().join("meta"),
+        &serde_json::to_vec(&meta)?,
+    )?;
+
+    log!(
+        "{} ({}, {}) -> {}",
+        style(&filename).dim(),
+        style(kind).yellow(),
+        style(arch).yellow(),
+        style(new_filename.display()).cyan(),
+    );
+    let mut out = fs::File::create(&new_filename)?;
+
+    if compression_level > 0 {
+        copy_encode(writable_data.as_slice(), &mut out, compression_level)?;
+    } else {
+        io::copy(&mut writable_data.as_slice(), &mut out)?;
+    }
+    rv.push((get_dyld_image_unified_id(&in_object), kind));
 
     Ok(rv)
 }
@@ -439,56 +330,15 @@ fn sort_files(sort_config: &SortConfig, paths: Vec<PathBuf>) -> Result<(usize, u
                 }
             // dyld shared cache
             } else if let Ok(cache) = DyldCache::<Endianness>::parse(bv.as_slice()) {
-                // let image_debug_ids = cache
-                //     .images()
-                //     .filter_map(|image| {
-                //         let image_offset = image.file_offset().ok()? as usize;
-                //         let name = image.path().ok()?.rsplit('/').next().unwrap().to_string();
-                //         let img_bv = ByteView::from_slice(&bv[image_offset..]);
-                //         if Archive::peek(&img_bv) != FileFormat::Unknown {
-                //             let in_object = image.parse_object().ok()?;
-                //             Some(
-                //                 process_dyld_file(sort_config, img_bv, in_object, name.clone())
-                //                     .map_err(|err| {
-                //                         eprintln!(
-                //                             "{}: ignored error {} ({})",
-                //                             style("error").red().bold(),
-                //                             err,
-                //                             style(name).cyan(),
-                //                         )
-                //                     })
-                //                     .ok()?
-                //                     .into_iter()
-                //                     .map(|x| x.0),
-                //             )
-                //         } else {
-                //             eprintln!(
-                //                 "{}: ignored error {} ({})",
-                //                 style("error").red().bold(),
-                //                 anyhow!("unrecognized file"),
-                //                 style(name).cyan(),
-                //             );
-                //             None
-                //         }
-                //     })
-                //     .flatten();
-
-                // debug_ids.lock().unwrap().extend(image_debug_ids);
-
                 for image in cache.images() {
-                    let image_offset = image.file_offset()? as usize;
-                    let img_bv = ByteView::from_slice(&bv[image_offset..]);
-                    if Archive::peek(&img_bv) != FileFormat::Unknown {
-                        let name = image.path()?.rsplit('/').next().unwrap().to_string();
-                        let in_object = image.parse_object()?;
-                        debug_ids.lock().unwrap().extend(
-                            process_dyld_file(sort_config, img_bv, in_object, name)?
-                                .into_iter()
-                                .map(|x| x.0),
-                        );
-                    }
+                    // let in_object = image.parse_object()?;
+                    debug_ids.lock().unwrap().extend(
+                        process_dyld_image(sort_config, image)?
+                            .into_iter()
+                            .map(|x| x.0),
+                    );
                 }
-                // object file directly
+            // object file directly
             } else if Archive::peek(&bv) != FileFormat::Unknown {
                 for (unified_id, object_kind) in process_file(
                     sort_config,
